backend/requirements.txt
Django==4.2.5
gunicorn==20.1.0
psycopg2-binary==2.9.6
APScheduler==3.10.1
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
django-apscheduler==0.6.2
django-filter==23.2
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.5
fonttools==4.39.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
networkx==3.1
numpy==1.24.3
openapi-codec==1.3.2
packaging==23.1
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
scikit-fuzzy==0.4.2
scipy==1.10.1
setuptools==57.0.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==2.0.2
wheel==0.36.2
django-cors-headers==4.0.0